Refactor code
Extracted the parameters of the `fillOutRegisterForm` method into a model object to represent it as an Entity. This allows us to add more meaning to our code by reducing the amount of parameters of the method and relating them to something meaningful. Also, I separated the code with break lines to tell at a glance which are the "Arrange, Act and Assert" parts of the code in each test method.